Exponentiating objects (and the functor they induce), along with an experimental redesign of the Exponential type.
The experiment was to remove the requirement that 'A' have products with all other objects.  It wasn't as painful as expected, but I'm still not convinced it's worth the pain it does cause.

The Functor proof in Categories.Object.Exponentiating needs a lot of cleanup.  I finally made the right mental connections to be able to grok a proof and get it to compile, next step will be massive cleanup.